rename Text children components
